Add dot.
