+++ src/main/java/org/rumbledb/item/parsing/ItemParser.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.rumbledb.itemparsing;
+package org.rumbledb.item.parsing;
 
 import com.jsoniter.JsonIterator;
 import com.jsoniter.ValueType;
