+++ src/main/java/org/rumbledb/item/parsing/JSONSyntaxToItemMapper.java
@@ -18,24 +18,42 @@
  *
  */
 
-package org.rumbledb.itemparsing;
+package org.rumbledb.item.parsing;
 
-import org.apache.spark.api.java.function.Function;
-import org.apache.spark.sql.Row;
+import com.jsoniter.JsonIterator;
+import org.apache.spark.api.java.function.FlatMapFunction;
 import org.rumbledb.api.Item;
 import sparksoniq.jsoniq.runtime.metadata.IteratorMetadata;
 
-public class RowToItemMapper implements Function<Row, Item> {
+import java.util.Iterator;
+
+public class JSONSyntaxToItemMapper implements FlatMapFunction<Iterator<String>, Item> {
 
     private static final long serialVersionUID = 1L;
     private final IteratorMetadata metadata;
 
-    public RowToItemMapper(IteratorMetadata metadata) {
+    public JSONSyntaxToItemMapper(IteratorMetadata metadata) {
         this.metadata = metadata;
     }
 
     @Override
-    public Item call(Row row) throws Exception {
-        return ItemParser.getItemFromRow(row, metadata);
+    public Iterator<Item> call(Iterator<String> stringIterator) throws Exception {
+        return new Iterator<Item>() {
+            @Override
+            public boolean hasNext() {
+                return stringIterator.hasNext();
+            }
+
+            @Override
+            public Item next() {
+                JsonIterator object = JsonIterator.parse(stringIterator.next());
+                return ItemParser.getItemFromObject(object, metadata);
+            }
+
+            @Override
+            public void remove() {
+                throw new UnsupportedOperationException();
+            }
+        };
     }
 }
