+++ src/main/java/sparksoniq/semantics/DynamicContext.java
@@ -28,7 +28,7 @@ import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.rumbledb.api.Item;
-import org.rumbledb.itemparsing.RowToItemMapper;
+import org.rumbledb.item.parsing.RowToItemMapper;
 
 import sparksoniq.exceptions.OurBadException;
 import sparksoniq.exceptions.SparksoniqRuntimeException;
