+++ src/main/java/sparksoniq/spark/iterator/function/JsonFileFunctionIterator.java
@@ -22,7 +22,7 @@ package sparksoniq.spark.iterator.function;
 
 import org.apache.spark.api.java.JavaRDD;
 import org.rumbledb.api.Item;
-import org.rumbledb.itemparsing.JSONSyntaxToItemMapper;
+import org.rumbledb.item.parsing.JSONSyntaxToItemMapper;
 
 import sparksoniq.exceptions.CannotRetrieveResourceException;
 import sparksoniq.jsoniq.ExecutionMode;
